Add extraDependencies
Supposed to have the same effect it has in nixpkgs: included in the
runtime closure but not visible to the user.

This is useful for preventing GC of build dependencies that would not
otherwise be included in the runtime closure but are likely to be needed
again in the future.