Keep all fish-completions packages around
...even empty ones.

For packages with no Fish completions in `home.packages`, we build an
empty directory that is not included in our runtime closure (because we
build a collection of symlinks to their contents: if there are no
contents, there is nothing to symlink). This means these empty
completion packages get garbage-collected by Nix, and although they are
not that expensive to rebuild there can be enough of them it adds up.

Fix this by forcing a dependency on all of them. Since we already
depended on them anyway if they were non-empty, this only adds
dependencies on empty directories (so should not significantly affect
storage space).