nushell: protect plugin dependencies from nix gc
plugin.msgpackz contains references to store paths, but it is a
compressed file so Nix does not automatically detect them. Since they
will typically not be referenced elsewhere, this means the plugins get
garbage-collected while nushell is still configured to use them.

Fix this by including a separate plain-text file with the plugin store
paths for Nix to find.

Fixes #7070